Fix volume increase on track change or restart
When clicking on the currently playing item in "Now Playing" a few
times:

In Chrome, an exception is thrown:

Failed to execute 'createMediaElementSource' on 'AudioContext': HTMLMediaElement already connected previously to a different MediaElementSourceNode.

but Firefox is happy to try connecting the same source multiple times,
leading to apparent volume increases despite gain and volume settings
remaining constant.